# resources/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KK Wholesalers ERP — Route & API Documentation
  version: 1.0.0
  description: |
    Comprehensive route and API documentation for **KK Wholesalers Monolithic ERP & Inventory Management System**.
    
    ### Key Highlights
    - **Concurrency-Safe Stock Ledger**: Pessimistic database locking (`lockForUpdate`) on `store_stocks` and immutable double-entry `stock_movements`.
    - **Role-Based Access Control (RBAC)**: Administrator, Branch Manager, and Store Manager scopes.
    - **POS Checkout & Invoicing**: Real-time stock validation, multi-item checkout, sequential invoice numbering (`INV-YYYYMMDD-XXXX`).
    - **Inter-Store Transfers**: Origin store balance deduction and destination store balance addition with dispatch transfer notes (`TRF-XXXXXXXX`).
    - **Monitoring & Observability**: Multi-channel daily structured logging integrated with Grafana, Loki, and Promtail.
servers:
  - url: http://localhost:8000
    description: Local Development Server (Artisan)
  - url: http://localhost:9000
    description: Docker Nginx Web Gateway
...
paths:
  /login:
    get:
      tags:
        - Authentication & Session
      summary: Display Login Page
      description: Returns the Blade login view. Redirects to `/dashboard` if already authenticated.
      responses:
        '200':
          description: Login form rendered successfully
        '302':
          description: Authenticated user redirected to dashboard
    post:
      tags:
        - Authentication & Session
      summary: Authenticate User
      description: Validates user credentials, authenticates session, regenerates CSRF token, and logs authentication events.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: admin@kkwholesalers.com
                password:
                  type: string
                  format: password
                  example: password123
                remember:
                  type: boolean
                  example: false
      responses:
        '302':
          description: Successful authentication redirecting to `/dashboard`
        '422':
          description: Invalid credentials or validation failure
...
  /dashboard:
    get:
      tags:
        - Executive Dashboard
      summary: View Executive Dashboard
      description: Returns sales metrics, total retail/cost stock valuation, low-stock warnings, and store leaderboards scoped to user's permissions.
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Dashboard metrics rendered successfully
        '302':
          description: Unauthenticated redirect to `/login`
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: laravel_session
      description: Standard Laravel session cookie authentication with CSRF protection (`X-CSRF-TOKEN`).
...